challenges/web-kinjutsu/src/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
